+++ CHANGES.rst
@@ -7,6 +7,7 @@ Version 3.7.1 (2020-02-XX)
 
 * GH227 Fix a Type error when loading categorical data in dask without
   specifying it explicitly
+* No longer trigger the SettingWithCopyWarning when using bucketing
 
 Version 3.7.0 (2020-02-12)
 ==========================